feat: extract out fetchVendorSettings
