+++ src/routes/tsx/routes.tsx
@@ -34,7 +34,6 @@ import { sendResetPassword, sendLink, sendCode } from "../../controllers/email";
 import { validateCode } from "../../authentication-flows/passwordless";
 import { getUsersByEmail } from "../../utils/users";
 import userIdGenerate from "../../utils/userIdGenerate";
-import { vendorSettingsSchema } from "../../types";
 import { sendEmailVerificationEmail } from "../../authentication-flows/passwordless";
 import { getSendParamFromAuth0ClientHeader } from "../../utils/getSendParamFromAuth0ClientHeader";
 import {
@@ -44,23 +43,7 @@ import {
 } from "../../utils/authCookies";
 import { setCookie, getCookie } from "hono/cookie";
 import { waitUntil } from "../../utils/wait-until";
-
-const DEFAULT_SESAMY_VENDOR = {
-  name: "sesamy",
-  logoUrl: `https://assets.sesamy.com/static/images/email/sesamy-logo.png`,
-  style: {
-    primaryColor: "#7D68F4",
-    buttonTextColor: "#7D68F4",
-    primaryHoverColor: "#7D68F4",
-  },
-  loginBackgroundImage: "",
-  checkoutHideSocial: false,
-  supportEmail: "support@sesamy.com",
-  supportUrl: "https://support.sesamy.com",
-  siteUrl: "https://sesamy.com",
-  termsAndConditionsUrl: "https://store.sesamy.com/pages/terms-of-service",
-  manageSubscriptionsUrl: "https://account.sesamy.com/manage-subscriptions",
-};
+import { fetchVendorSettings } from "../../utils/fetchVendorSettings";
 
 async function initJSXRoute(state: string, env: Env) {
   const session = await env.data.universalLoginSessions.get(state);
@@ -87,33 +70,6 @@ async function initJSXRoute(state: string, env: Env) {
   initI18n(tenant.language || "sv");
 
   return { vendorSettings, client, tenant, session };
-}
-
-async function fetchVendorSettings(
-  env: Env,
-  client_id?: string,
-  vendor_id?: string,
-) {
-  if (!vendor_id && !client_id) {
-    return DEFAULT_SESAMY_VENDOR;
-  }
-
-  const vendorId = vendor_id || client_id;
-
-  try {
-    const vendorSettingsRes = await fetch(
-      `${env.API_URL}/profile/vendors/${vendorId}/style`,
-    );
-
-    const vendorSettingsRaw = await vendorSettingsRes.json();
-
-    const vendorSettings = vendorSettingsSchema.parse(vendorSettingsRaw);
-
-    return vendorSettings;
-  } catch (e) {
-    console.error(e);
-    return DEFAULT_SESAMY_VENDOR;
-  }
 }
 
 // duplicated from /passwordless route
