+++ src/utils/fetchVendorSettings.ts
@@ -1,0 +1,45 @@
+import { vendorSettingsSchema, Env } from "../types";
+
+const DEFAULT_SESAMY_VENDOR = {
+  name: "sesamy",
+  logoUrl: `https://assets.sesamy.com/static/images/email/sesamy-logo.png`,
+  style: {
+    primaryColor: "#7D68F4",
+    buttonTextColor: "#7D68F4",
+    primaryHoverColor: "#7D68F4",
+  },
+  loginBackgroundImage: "",
+  checkoutHideSocial: false,
+  supportEmail: "support@sesamy.com",
+  supportUrl: "https://support.sesamy.com",
+  siteUrl: "https://sesamy.com",
+  termsAndConditionsUrl: "https://store.sesamy.com/pages/terms-of-service",
+  manageSubscriptionsUrl: "https://account.sesamy.com/manage-subscriptions",
+};
+
+export async function fetchVendorSettings(
+  env: Env,
+  client_id?: string,
+  vendor_id?: string,
+) {
+  if (!vendor_id && !client_id) {
+    return DEFAULT_SESAMY_VENDOR;
+  }
+
+  const vendorId = vendor_id || client_id;
+
+  try {
+    const vendorSettingsRes = await fetch(
+      `${env.API_URL}/profile/vendors/${vendorId}/style`,
+    );
+
+    const vendorSettingsRaw = await vendorSettingsRes.json();
+
+    const vendorSettings = vendorSettingsSchema.parse(vendorSettingsRaw);
+
+    return vendorSettings;
+  } catch (e) {
+    console.error(e);
+    return DEFAULT_SESAMY_VENDOR;
+  }
+}
